Implement interface org.hipparchus.CalculusFieldElement for IExpr
See:
- https://github.com/Hipparchus-Math/hipparchus/issues/138

+++ symja_android_library/matheclipse-core/src/main/java/org/matheclipse/core/interfaces/IComplex.java
@@ -42,7 +42,7 @@ public interface IComplex extends IBigNumber {
    * @return <code>this<sup>exp</sup></code>
    * @throws ArithmeticException if {@code 0^0} is given.
    */
-  public IComplex pow(long n);
+  public IComplex pow(int n);
 
   public IComplex[] quotientRemainder(final IComplex c2);
